Feat: GlobalLoadingBar 컴포넌트에서 로딩 여부를 판단할 커스텀훅 추가
